Renovating ServerLogStorage to help with concurrency.
Break the ActivityLogMetadata and ActivityLogFileAppender into a pipe of small parts that each do simple things.
Each ILogEntryProcessor in the chain can assume:
- They have the same lifetime
- They are operating single threaded (per ServerTask/shard)
- They won't be disposed until they are completed
- They won't be evicted from the cache until they are completed

Replace the implicit coupling of SensitiveValues through DeploymentTaskController and the Log-tree with explicit coupling to the MaskSensitiveVariables Processor.

Also added a Retry for ObjectDisposedException, IOException and TimeoutException to help us be resilient to Transient failures that could be overcome by simply trying again.