We need a custom ProxyCache because of how we use it.
Most of the changes here were related to upgrading from NuGet 3.4.3 -> 3.5.0-octopus-*.

The other change was to unwind some complexity in how we configure SourceRepository where we were losing credentials and then needing to find them again. This version of NuGet looks for credentials on the Source itself before falling back to the CredentialCache.