Auto-save when calling the setter in settings.
There is a lot of potentially confusing code around these days, calling `.Save()` on an abstraction, which actually encapsulates the same instance of the AbstractKeyValueStore. There are also cases where we need to call save, so another completely disconnected piece can get its hands on the saved information.

I think it is better if we just auto-save after every call to Set a value. This way developers can't make mistakes like forgetting to call `Save()` at the right time.