Moving CancellationToken to call path.
It was really starting to smell having the only way to set the `CancellationToken` being the constructor, then forcing us to pass around the same instance. The problem is revealed where we had to Initialize() the `ThrottledClusterWideMutex` configuring a static `Throttle` instance... I think this will scale better if we can acquire a `ThrottledClusterWideMutex` which may possibly be a singleton, and hand it the appropriate CancellationToken for the consumer.

Now I just need to figure out how to best update the calling code. :(