Make EventStore SingleInstance for performance
EventStore was InstancePerDependency, but also creates an expensive cryptography dependency in its constructor. This appears to be used as a way to emulate a one-way hash using the `MasterKeyEncryption`.

I've changed this to be a one-way hash using SHA1.

I also found a lot of places where we were not disposing of certain crypto instances.