Stop double logging errors in ConsoleHost
At some point in the past, we realised certain exception details (including the ExceptionKnowledgeBase) were only being written to Console.Out, but not the log files. Over time we've been emitting that information to the log files, but neglected to remove it from the ConsoleHost - end result is we've been logging the same things multiple times. For example, if the database upgrade fails, the same error message is logged ~9 times at different points. This change should alleviate some of that noise.