Set home directory when creating instance
The config file has to be in the root of the home directory by default. If we initialize the home folder at this point, we can load the instance immediately after this command has completed. This means all future logs will be written to the instance's log folder. The home folder can always be moved afterwards by using the configure command.