Less amgibious logging of the exe
Sometimes it's in the working directory, sometimes it's not. This logging would have helped figure out some bugs where the exe being called was completely unrelated to the working directory, but we had no idea.

I think this is a good compromise between the two approaches, since in most cases we'll get a nice short exe message when it's in the same working directory, otherwise we'll get the full path information.