Improved logging at program startup
I've taken a long-hard look at how we log startup diagnostics, and our command-line is really noisy. I've introduced a new NLog named logger so we can redirect these diagnostics to the best log file for the job, and leave the actual console output nice and clean and pristine. This change also removes the need for `nologo`. Finally I've gone through and improved how/when we load the current instance, and how we leverage the shared/global log file, and how we leverage the instance-specific log file.

For example: we'll log startup diagnostics to the global log file, then try the load the instance and log the same diagnostics to the instance's log file - this makes it easier to understand what's happening in a multi-instance situation.