Prevent breaking existing automation scripts
After discussion we decided it's not worth breaking anyone's automation scripts since the --nologo --noconsolelogging --console switches are all safe enough to ignore if they are provided on the args. The worse thing would be for someone to miss the warnings (which are really just noise in the log file) and upgrade only to find their scripts broken.