Added SqlServerAppLockClusterWideMutex
The driver behind this was to acquire mutually exclusive access to the SQL Database when performing schema upgrades. I started looking at using the existing `ClusterWideMutex` but this requires actual database schema - so we end up with a chicken and egg problem.

To overcome this I've gone and implemented a pure SQL Server application lock implementation, and added a bunch of extra testing to boot.

In addition to these changes, I've increased the wait message level to Info since it wasn't being displayed by default.