Added the ProcessOutputSource.Debug to the ScriptLog
We recently introduced the capability to stream `Debug` level logs back to Octopus Server. This is useful for streaming back details of the boostrapping process etc which should't really exist as Info or Error logs. Unfortunately we didn't realise this code alsoe needed to be updated because Tentacle does its own conversions.