Add v2 ExecuteScript overload to ITentacleClient
Add an overload for ITentacleClient.ExecuteScript that accepts a `StartScriptCommandV2`

Using this overload will ensure a V2 endpoint is used to execute the script, falling back to V1 if V2 is unsupported.

This is intended to reduce confusion from the existing method which takes a v3alpha version of the command.

[SC-69441]